% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albert.rapp.de-package.R
\docType{package}
\name{albert.rapp.de-package}
\alias{albert.rapp.de}
\alias{albert.rapp.de-package}
\title{albert.rapp.de: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: John Gavin \email{john.b.gavin@gmail.com}

}
\keyword{internal}
